// apps/api/src/email/email-events.ts
export const EMAIL_EVENTS = {
  send: 'email.send',
  received: 'email.received',
} as const;